wip: refactor install command
